refactor(metrics): move kluctl_project related metrics to a separate file
