+++ docs/metrics/v1alpha1/README.md
@@ -13,3 +13,4 @@ weight: 10
 The controller exports several metrics in the [OpenMetrics compatible format](https://github.com/OpenObservability/OpenMetrics/blob/main/specification/OpenMetrics.md). They can be scraped by all sorts of monitoring solutions (e.g. Prometheus) or stored in a database. Because the controller is based on [controller-runtime](https://github.com/kubernetes-sigs/controller-runtime), all the [default metrics](https://book.kubebuilder.io/reference/metrics-reference.html) as well as the following controller-specific custom metrics are exported:
 
 - [kluctldeployment_controller](kluctldeployment_controller.md)
+- [kluctl_project](kluctl_project.md)
